Initial commit of Ducky Bot
So far this bot has been trained with a first version of the ball chase, basic score, default middle stage, and custom middle stage phases.

The custom middle stage is being actively worked on to try to improve on the standard implementation given by the community made guide for rlgym_sim. I am creating and testing my own ideas for rewards to do this.